Move parse_id_or_slug_param util
